site: Set resource defaults for Exec
Though I normally dislike resource defaults I make an exception for
Exec. This sets the `path` attribute for every Exec resource to the
value of the `path` fact from Facter ensuring that you don't get
Puppet furiously complaining about unqualified paths to binaries.
In turn this makes it easier to relocate binaries without manifests
exploding on you because of hard coded locations.

Additionally when an exec fails it will now show what happened, the
output of running the command, instead of hiding it.