Use Puppet to bootstrap the system
Once we have Puppet we can actually already use `puppet apply` to
configure the rest of the system. This means we can get rid of all
the `install` and `gem` calls and instead use Puppet manifests to
do the rest.

Though this is a bit slower than a pure Bash approach we gain a lot
in features, like idempotency and maintainability.