aio_binstub: Create binstubs for AIO puppet binaries
The AIO puppet binaries are in /opt/puppetlabs/puppet/bin which means
they're not on your path by default. Instead of adding that whole
directory to our path which would also override system things like
gem and irb this defined type allows us to create binstubs that call
out to the Puppetlabs binary instead.

Using `prefix` a command can be prefixed with a string like `puppet`
so that we create a binstub named `puppet-irb` instead of `irb`.