pupa: Don't trigger prerun or postrun commands
Once we've configured Puppet every next call to `papply` will cause
it to trigger the `prerun` command but we don't want that until
we're done fully configuring the machine. As such, we set the
`prerun_command` and `postrun_command` explicitly to nothing to
ensure that whatever is set in `puppet.conf` at that point is
ignored.