Don't rely/use prerun_command
When a Puppet run executes the `prerun_command` is only executed
after manifests are loaded and databinding lookups have taken place.

This means we can't depend on it to be run _before_ to deploy our
environment and would result in confusing behaviour and the need to
run Puppet twice in order to actually get our changes.