Restructure role::base
It was getting rather crowded in role::base and it shouldn't. It's
main purpose is to include modules (profiles are modules too) that
we need to apply to every machine, not contain actual resources
itself.

All resources were split up into the `pupa` profile which contains
the business logic of setting up the other necessities around Pupa
on this machine.

Additionally we create a `papply` script now that significantly
reduces the amount of characters we need to type to do a full
`puppet apply` run.