papply: Add locking, don't require environment
In order to prevent Puppet from running twice papply now writes and
checks for a lock file.

Additionally we no longer require the environment to be passed in.
This check is pretty naive, if the first argument starts with a `-`
we assume that the rest of the arguments will also be dash arguments
and not contain an environment name. If it doesn't start with a `-`
we assume it is the environment name and then shift the rest.