hiera/ssh: Add diffie-hellman-group14-sha1
Ubuntu 14.04 ships with an OpenSSH version that does not support
curve25519-sha256@libssh.org. Because of all the projects that
use libssh which does not support diffie-hellman-group-exchange-sha25
we would not be able to agree on a key exchange algorithm without
supporting the older diffie-hellman-group14-sha1.

The diffie-hellman-group14-sha1 is the 2048 bit group and at least
better than diffie-hellman-group1-sha1 which is the other supported
group but is only 768 bits.