Upgrade to Puppet 5
This commit is a large one, mainly because it removes all my
configuration. It's been useful for people to be able to fork and work
of this repository but due to it also containing my configs rebasing
against it to get updates has sucked. This fixes that.

This will now bootstrap a masterless/apply Puppet Platform 5
installation, with the usual goodies of Hiera EYAML and the nuking of
mcollective. It also ships with a sensible `puppet.conf` and a fully
default `hiera.yaml` for the production environment. It additionally
nukes the global `hiera.yaml` to avoid any nasty surprises. It's now
advised to only use environment and module level Hiera hierarchies and
only use a global one if needed to temporarily override a lookup
hierarchy.

The supplied `Puppetfile` now only includes a few useful libraries such
as stdlib, augeas shellvar and hash2stuff. Hash2stuff allows you to dump
Puppet hashes to INI, key/value (shell), JSON or YAML without needing
to resort to `[inline_]template()`.